arch/x86: Add 0x3e prefix to prevent CET violation
With Intel CET (especially for IBT), each branch target should start
with the ENDBR instruction.  But we cannot add that for dynamic
tracing as it creates artificially indirect branches in the middle of
the function.  Instead we can add the NO-TRACK prefix (0x3e) to
disable IBT for that branch.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>